backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
PyJWT==2.8.0
bcrypt==4.1.2 